refactor: reorganize shared app
