+++ app/registry/models/document.py
@@ -9,7 +9,7 @@ from django.contrib.contenttypes.models import ContentType
 from django.db import models
 
 from app.shared.constants import DocumentType, StatusDocument
-from app.shared.mixins import StatusableMixin, StatusHistory
+from app.shared.status.mixins import StatusHistory, StatusableMixin
 
 
 class Document(StatusableMixin, models.Model):
