+++ app/registry/models/registration.py
@@ -8,7 +8,7 @@ from django.dispatch import receiver
 from django.utils import timezone
 
 from app.shared.constants import StatusRegistration
-from app.shared.mixins import StatusableMixin
+from app.shared.status.mixins import StatusableMixin
 from app.timetable.models import Reservation
